Add files via upload
This C++ program functions as a simple vending machine, allowing users to select and purchase drinks, snacks, or candies based on a displayed menu. Users input their choice of item and the amount of money they have, and the program checks if they have sufficient funds to complete the transaction. It utilizes nested switch statements to manage different categories of items and their prices, providing feedback by confirming purchases with any change or indicating insufficient funds. Overall, the program facilitates a structured interaction for users to make selections and process payments.